requirements.txt
aladhan-api==6.1.0
APScheduler==3.10.4
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
oauthlib==3.2.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
salat==1.0.1
schedule==1.2.1
six==1.16.0
tweepy==4.14.0
tzlocal==5.2
urllib3==2.2.0
